requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.2
click==8.3.1
cryptography==46.0.3
dataclasses-json==0.6.7
datasets==2.16.1
decorator==4.4.2
dill==0.3.7
distro==1.9.0
dnspython==2.7.0
docx2txt==0.8
environs==9.5.0
faiss-cpu==1.13.1
fastapi==0.110.0
ffmpeg-python==0.2.0
filelock==3.18.0
fire==0.7.0
fonttools==4.56.0
frozenlist==1.8.0
fsspec==2023.10.0
future==1.0.0
grpcio==1.58.0
h11==0.14.0
hf-xet==1.1.2
httpcore==1.0.7
httptools==0.7.1
httpx==0.27.0
huggingface-hub==0.32.0
idna==3.10
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.1
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.53
langchain-text-splitters==0.0.1
langsmith==0.1.147
llvmlite==0.46.0
loguru==0.7.2
lxml==5.3.1
MarkupSafe==3.0.2
marshmallow==3.26.1
minio==7.2.20
more-itertools==10.8.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.15
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.2
numba==0.63.1
numpy==1.26.4
openai==1.68.2
openai-whisper==20231117
opencv-python==4.9.0.80
opencv-python-headless==4.11.0.86
orjson==3.11.5
packaging==23.2
pandas==2.2.1
pdf2docx==0.5.8
pdfminer.six==20231228
pillow==10.2.0
pluggy==1.6.0
proglog==0.1.12
propcache==0.4.1
protobuf==6.33.2
pyarrow==22.0.0
pyarrow-hotfix==0.7
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
pymilvus==2.3.5
pymongo==4.11.3
PyMuPDF==1.25.4
pypdf==4.2.0
pytesseract==0.3.10
pytest==8.1.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==2.3.1
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.10.1
SQLAlchemy==2.0.29
starlette==0.36.3
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.5.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.15.1
torch==2.7.0
tqdm==4.66.2
transformers==4.36.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
ujson==5.11.0
urllib3==2.4.0
uvicorn==0.27.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
